Save features working (last result + motion vector) - They open the data in a new tab.. Easily save from there...

+++ backend/backend/views.py
@@ -1,6 +1,7 @@
 import pprint
 
 from django.http import HttpResponse, FileResponse, JsonResponse
+from django.core.files.uploadedfile import TemporaryUploadedFile
 
 from django.views.decorators.csrf import csrf_exempt
 import os
@@ -13,7 +14,6 @@ from backend.settings import BASE_DIR
 import random
 
 from alg.image_api import Gui as Cache, open_series
-
 
 # cache = None
 # example = "EmekRefaim"
@@ -29,6 +29,7 @@ from alg.image_api import Gui as Cache, open_series
 # cache.setup(series_path=p, suffix=suffix, extension="jpg", height=500, width=900)
 cache = None
 
+
 def test(request):
     return HttpResponse("<html><body>Reached test!</body></html>")
 
@@ -36,29 +37,48 @@ def test(request):
 @csrf_exempt
 def upload_images(request):
     global cache
-    file = request.FILES["images"]
+    file: TemporaryUploadedFile = request.FILES["images"]
     try:
-        if not file:
-            raise ValueError
+        if not file or 'zip' not in file.content_type:
+            raise ValueError("Must supply valid zip file.")
+        name = file.name.replace(".zip", "").replace(" ", "_").strip()
         shutil.rmtree(settings.IMAGES_DIR)
         with zipfile.ZipFile(file, 'r') as zip_ref:
             zip_ref.extractall(settings.IMAGES_DIR)
         path = os.path.join(settings.IMAGES_DIR, os.listdir(settings.IMAGES_DIR)[0])
-        file_list =  os.listdir(path)
-        file_list.sort()
-        file_format = "{:03d}.jpg" if len(file_list)> 100 else "{:02d}.jpg"
-        for count, filename in enumerate(file_list):
-            os.rename(os.path.join(path, filename), os.path.join(path, file_format.format(count+1)))
-        cache = Cache()
-        cache.setup(series_path=path)
-        return JsonResponse({'rows': cache._rows,'cols':  cache._cols,'channels': cache._channels,'frames': cache._frames})
+        # Zip provided may contain files only, no dirs
+        if not os.path.isdir(path):
+            path = settings.IMAGES_DIR
+        file_list = os.listdir(path)
+        extension = os.path.splitext(file_list[0])[1].replace(".", "")
+        if extension.lower() not in ['jpg', 'jpeg', 'png']:
+            raise ValueError("Images' format must be jpeg or png, not {}".format(extension))
+        image_only_list = list(filter(lambda p: p.endswith(extension), file_list))
+        irrelevant_file_list = list(filter(lambda p: not p.endswith(extension), file_list))
+        if len(image_only_list) > 999:
+            raise ValueError("Up to 999 images supported, received {}".format(len(image_only_list)))
+        image_only_list.sort()
+        file_format = "{}{:03d}.{}" if len(image_only_list) > 100 else "{}{:02d}.{}"
+        for count, filename in enumerate(image_only_list):
+            os.rename(os.path.join(path, filename), os.path.join(path, file_format.format(name, count + 1, extension)))
+        for filename in irrelevant_file_list:
+            try:
+                os.remove(os.path.join(path, filename))
+            except:
+                print("Could not remove {}".format(filename))
+        cache = Cache(series_path=path, prefix=name, extension=extension)
+        return JsonResponse({'rows': cache._rows,
+                             'cols': cache._cols,
+                             'channels': cache._channels,
+                             'frames': cache._frames
+                             })
     # load images to directory
 
-    except Exception as e :
+    except Exception as e:
+        print(e)
         response = HttpResponse(str(e))
         response.status_code = 400
         return response
-    return HttpResponse('')
 
 
 @csrf_exempt
@@ -78,11 +98,13 @@ def slice(request):
 def focus(request):
     res = cache.get_last_result()
     try:
-        depth = request.GET.get('depth')
+        depth_input = request.GET.get('depth')
         center = int(request.GET.get('center'))
         radius = int(request.GET.get('radius'))
-        depth = float(depth)
+        depth = 1 - float(depth_input)
         # path = os.path.join(BASE_DIR, 'sample_data', 'apples', 'APPLE{:03d}.jpg'.format(int(value * 200)))
+        assert 1 <= center <= cache._frames
+        assert 1 <= radius <= cache._frames / 2
 
         res = cache.focus(depth, center, radius)
     except Exception as e:
@@ -97,19 +119,22 @@ def viewpoint(request):
         res: np.ndarray = cache.get_last_result()
         try:
             slice_raw = request.GET.get('slice')
+            factor = request.GET.get('factor', None)
+            if factor:
+                factor = float(factor)
             if slice_raw not in [None, "", "()", "((),())", (), []]:
                 print("Viewpoint - slice {}".format(slice_raw))
                 inputs = [int(i) for i in slice_raw.split(",")]
                 slice = (inputs[0], inputs[1]), (inputs[2], inputs[3])
 
-                res = cache.viewpoint(slice=slice)
+                res = cache.viewpoint(slice=slice, factor=factor)
             else:
                 shift = float(request.GET.get('shift'))
                 move = float(request.GET.get('move'))
                 stereo = float(request.GET.get('stereo'))
                 print("Viewpoint - move: {} stereo: {} shift: {}".format(move, stereo, shift))
 
-                res = cache.viewpoint(shift=shift, move=move, stereo=stereo)
+                res = cache.viewpoint(shift=shift, move=move, stereo=stereo, factor=factor)
         except Exception as e:
             print(e)
